Fix patchConstant when the user does not specify $class_method
When a user calls patchConstant without the argument $class_method the docs explain the constant will changed everywhere.

However, the code will store in $patches_to_apply[$constant] the result of strtolower($class_name) ==> strtolower(null) ==> "". When we check if we have to return the patched value we will always return false, as we check with `isset` the variable $patches_to_apply[$constant] (this will always be true) and when we proceed to check the value of $patches_to_apply[$constant] (the empty string) with the current class it always fails the test.

The end result is that a call to patchConstant('FOO', 1); never changes the constant FOO.